+++ lib/go_http.dart
@@ -33,6 +33,7 @@ export 'src/hsts_cache.dart';
 export 'src/interceptors/auth_interceptor.dart';
 export 'src/interceptors/interceptor.dart';
 export 'src/interceptors/logging_interceptor.dart';
+export 'src/json_helpers.dart';
 export 'src/limits.dart';
 export 'src/logger.dart';
 export 'src/metrics/console_metrics_sink.dart';